+++ R/amps.R
@@ -190,6 +190,7 @@ readamps_d1ice <- function(date, time.resolution = "daily", xylim = NULL,
 #' @examples
 #' af <- amps_d1files()
 #' w <- readamps_d1wind(latest = TRUE, inputfiles = af)
+#' #w <- readamps_d2wind(latest = TRUE)
 #' vlen <- function(a, b) sqrt(a * a + b * b)
 #' plot(vlen(w[[1]], w[[2]]))
 #'
@@ -200,7 +201,7 @@ readamps_d1ice <- function(date, time.resolution = "daily", xylim = NULL,
 #'  cr1[,2] <- cr1[,2] + values(x[[2]]*scale);
 #'  segments(cr[sub,1], cr[sub,2], cr1[sub,1], cr1[sub,2])
 #' }
-#' arr(w, sample(ncell(w), 10000), scale = 1000)
+#' arr(w, sample(ncell(w), 10000), scale = 15000)
 #' @export
 readamps_d1wind <- function(date, time.resolution = "4hourly", xylim = NULL,
                             magonly = FALSE, dironly = FALSE, uonly = FALSE, vonly = FALSE,
@@ -298,6 +299,106 @@ readamps_d1wind <- function(date, time.resolution = "4hourly", xylim = NULL,
 
   r
 }
+
+#' @export
+readamps_d2wind <- function(date, time.resolution = "4hourly", xylim = NULL,
+                            magonly = FALSE, dironly = FALSE, uonly = FALSE, vonly = FALSE,
+                            latest = TRUE, returnfiles = FALSE, level = 1, ..., inputfiles = NULL) {
+
+  time.resolution <- match.arg(time.resolution)
+  if ((magonly + dironly + uonly + vonly) > 1) stop("only one of magonly, dironly, uonly or vonly may be used, exiting")
+
+  if (is.null(inputfiles)) {
+    files <- amps_d2files(time.resolution = time.resolution)
+  } else {
+    files <- inputfiles
+  }
+  if (returnfiles) return(files)
+
+  if (missing(date)) date <- if (latest) max(files$date) else min(files$date)
+  date <- timedateFrom(date)
+  ## findex <- .processDates(date, files$date, time.resolution)
+  files <- .processFiles(date, files, time.resolution)
+
+
+  nfiles <- nrow(files)
+  if (nfiles > 1L & !magonly & !dironly & !uonly & !vonly) {
+    warning("only one time index can be read at once unless 'magonly', 'dironly', 'uonly' or 'vonly' is TRUE")
+    files <- files[1L,]
+    nfiles <- 1L
+  }
+
+
+  if (!(magonly | dironly))
+    rasterfun <- function(x1, x2) {
+      x <- brick(x1, x2)
+      names(x) <- c("U", "V")
+      x
+    }
+  if (magonly)
+    rasterfun <- function(x1, x2) sqrt(x1 * x1 + x2 * x2)
+  if (dironly)
+    rasterfun <- function(x1, x2) (90 - atan2(x2, x1) * 180/pi)%%360
+
+  if (!(magonly | dironly)) {
+    if (uonly) rasterfun <- function(x1, x2) x1
+    if (vonly) rasterfun <- function(x1, x2) x2
+  }
+
+
+  cropit <- FALSE
+  if (!is.null(xylim)) {
+    cropit <- TRUE
+    cropext <- extent(xylim)
+  }
+
+  r <- vector("list", nfiles)
+
+  is_first_hour <- grepl("f000", basename(files$fullname))
+
+  ## we can't hardcode the bands since 202010 ...
+  bands <- c(5, 27)
+  ## so if total bands is > 254 we use this number
+  if (vapour::vapour_raster_info(files$fullname[ifile])$bands > 254) {
+    bands <- c(33, 40)
+  }
+  for (ifile in seq_len(nfiles)) {
+    r1 <- readwrf0(files$fullname[ifile], band = bands[1L] + is_first_hour[ifile] * 4) #raster(files$ufullname[ifile], band = files$band[ifile])
+    r2 <- readwrf0(files$fullname[ifile], band = bands[2L] + is_first_hour[ifile] * 4) #raster(files$vfullname[ifile], band = files$band[ifile])
+    r0 <- rasterfun(r1, r2)
+    #if (lon180)     r0 <- suppressWarnings(.rotate(r0))
+    if (cropit) r0 <- crop(r0, cropext)
+    r[[ifile]] <- r0
+
+  }
+  r <- stack(r)
+  if (magonly | dironly | uonly | vonly)  {
+    r <- setZ(r, files$date)
+    names(r) <- sprintf("wind_%s", format(files$date, "%Y%m%d"))
+  } else {
+
+    r <- setZ(r, rep(files$date, 2L))
+    names(r) <- sprintf("%swind_%s", c("U", "V"), format(files$date, "%Y%m%d"))
+  }
+
+
+  ## get alignment right (put this in raster?)
+  extent(r) <- extent(c(xmin(r) + res(r)[1]/2, xmax(r) + res(r)[1]/2,
+                        ymin(r), ymax(r)))
+
+  ##if (is.na(projection(r))) projection(r) <- "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
+
+  ## need to determine if "filename" was passed in
+  dots <- list(...)
+  if ("filename" %in% names(dots)) {
+    r <- writeRaster(r, ...)
+  }
+
+
+  r
+}
+
+